cleanup noya-keymap
