+++ packages/noya-keymap/src/index.ts
@@ -1,6 +1,5 @@
 export * from './types';
 export * from './Platform';
 export * from './useKeyCommands';
-// export * from './names';
 
-// export const FALLTHROUGH = 'fallthrough';
+export const FALLTHROUGH = 'fallthrough';
