--- packages/noya-keymap/src/prev/KeyCommands/index.ts
@@ -1,1 +1,0 @@
-export * from './KeyCommands';
